requirements.txt
#app
uvicorn==0.23.2
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi==0.104.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pyinstaller==6.2.0

#tests
pytest==7.4.3
pytest-mock==3.12.0
httpx==0.25.2